Whiskey Tango Foxtrot
Okay. It’s been a while. I’ve lost track of what I was doing with this.

But after a while trying to do it all on the Pi and having rotary
controller issues, I’m back to this.

We are about to turn this into a basic device where the controllers
display on the board. That’s it for now. Once that’s done I’ll put code
in to communicate the status to the pi during pauses.