v0.3.2 - Basics Working
Using 115200 as the baud rate was apparently bad despite being spec. I
couldn’t make this work but a sample thing was working using almost the
exact same technique. I pasted in a block that overwrote the baud to
the common 9600 and it worked! GAH!

Now I have the basics: the encoders transmit Vv/Cc for Volume or
Channel up or down and it takes a stream of 21 values: RGB each for
seven pixels.